Added default cell view
