--- macpaw/Views/CustomCellView.swift
@@ -1,20 +1,0 @@
-//
-//  CustomCellView.swift
-//  macpaw
-//
-//  Created by Andrew Shakula on 20.07.2022.
-//
-
-import SwiftUI
-
-struct CustomCellView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct CustomCellView_Previews: PreviewProvider {
-    static var previews: some View {
-        CustomCellView()
-    }
-}
